// src/Anfx.Pasivos.Frontend/src/api/models/pagoItemDto.ts
/**
 * Anfx.Pasivos.ApiService
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PagoItemDto { 
    idPago?: number;
    idContratoPasivo?: number;
    tipoPago?: string | null;
    cuentaBancaria?: string | null;
    fecPagoValor?: string | null;
    fecPagoRegistro?: string | null;
    montoPago?: number;
    montoAplicado?: number;
    montoAplicadoOtros?: number;
    saldoPago?: number;
}