// src/Anfx.Pasivos.Frontend/src/api/models/detalleMovimientoPagoDto.ts
/**
 * Anfx.Pasivos.ApiService
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DetalleMovimientoPagoDto { 
    idMovimiento?: number;
    fondeador?: string | null;
    contrato?: string | null;
    noPago?: number;
    descripcion?: string | null;
    fecMovimiento?: string | null;
    capitalPagado?: number;
    interesPagado?: number;
    ivaPagado?: number;
    totalPagado?: number;
    estatus?: boolean;
}