// src/Anfx.Pasivos.Frontend/src/api/models/autocompleteResultDto.ts
/**
 * Anfx.Pasivos.ApiService
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AutocompleteResultDto { 
    label?: string | null;
    value?: string | null;
}